Part 2 Completed
For part 2 I have chosen Pickup system. To implement this feature into asteroids game I created new class "Pickup". This new class has separate object type, additional interaction with spaceship and bullets. When colliding with spaceship pickup is destroyed and 2 things happen. First - one more life is added for player, so that he can survive longer and rewarding player for completing levels. Second - additional 50 is added to the score.

In addition to part 2, I changed slightly how score is added for destruction of the Asteroids and sub Asteroid, 30 and 10 respectively.